requirements.txt
newspaper3k
schedule
requests
lxml_html_clean
pillow
flask
